actions module for parsing hooks
Now the recognizer will call hooks back to signal shift/reduce/recursion
events through an 'Actions' module. This module holds a 'node' type for
parsing tree nodes, along with a comparison and constructor functions,
which the parser will use to discriminate different parsing trees (the
recogniser need not to do that).

* code cleanup
* bug fix: the "first" symbol set optimize search of newly activated
rules (the "predicted" item set)
* made debug messages really debug messages
* moved 'first' set instance creation into 'parse' function